// src/app/page/master/jenis-suku-bunga/jenis-suku-bunga.interface.ts
import {EventEmitter} from '@angular/core';

export interface JenisSukuBunga {
	namaJenisProduk?: string;//
	kdKelompokProduk?: number;//
	kdAccount?: number;
	kdJenisProduksiHead?: string;
	// kdDepartemen?: string;

	kode?: any;
	id?: any;
	kdProfile?: string;
	version?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
}